Core/DB: changed int to float

+++ sql/updates/characters/2022_03_22_00_Update_Characters.sql
@@ -1,6 +1,6 @@
 ALTER TABLE `characters` 
-ADD COLUMN `camp_x` int(10) NULL DEFAULT NULL AFTER `numRespecs`,
-ADD COLUMN `camp_y` int(10) NULL DEFAULT NULL AFTER `camp_x`,
-ADD COLUMN `camp_z` int(10) NULL DEFAULT NULL AFTER `camp_y`,
-ADD COLUMN `camp_o` int(10) NULL DEFAULT NULL AFTER `camp_z`,
-ADD COLUMN `camp_mapid` int(10) NULL DEFAULT NULL AFTER `camp_o`;
+ADD COLUMN `camp_x` float NULL DEFAULT NULL AFTER `numRespecs`,
+ADD COLUMN `camp_y` float NULL DEFAULT NULL AFTER `camp_x`,
+ADD COLUMN `camp_z` float NULL DEFAULT NULL AFTER `camp_y`,
+ADD COLUMN `camp_o` float NULL DEFAULT NULL AFTER `camp_z`,
+ADD COLUMN `camp_mapid` float NULL DEFAULT NULL AFTER `camp_o`;
